+++ vpin-connector-vps/src/main/java/de/mephisto/vpin/connectors/vps/matcher/TableMatcher.java
@@ -1,4 +1,4 @@
-package de.mephisto.vpin.server.vps;
+package de.mephisto.vpin.connectors.vps.matcher;
 
 import de.mephisto.vpin.connectors.vps.model.VpsAuthoredUrls;
 import de.mephisto.vpin.connectors.vps.model.VpsTable;
@@ -6,37 +6,34 @@ import de.mephisto.vpin.connectors.vps.model.VpsTable;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.text.similarity.CosineDistance;
 import org.apache.commons.text.similarity.EditDistance;
+import org.apache.commons.text.similarity.JaroWinklerDistance;
 
 import java.util.List;
 
 public class TableMatcher {
 
-  private EditDistance<Double> ed = new CosineDistance();
+  private EditDistance<Double> cd = new CosineDistance();
+  private EditDistance<Double> jwd = new JaroWinklerDistance();
 
-  private double THRESHOLD_NOTFOUND = 5.5;
+  private double THRESHOLD_NOTFOUND = 4.5;
+  private double THRESHOLD_CONFIRM = 0.3;
 
   //--------------------------------------
 
-  public VpsTable findClosest(String _fileName, String _rom, boolean checkall,
-      String _tableName, String _manuf, int _year, List<VpsTable> tables) {
+  public VpsTable findClosest(String fileName, String rom, String tableName, String manuf, int year, List<VpsTable> tables) {
 
-    final double[] minDistance = new double[]{10000};
+    final double[] minDistance = new double[] {10000};
     final VpsTable[] found = new VpsTable[1];
 
-    String romForFilter = StringUtils.isEmpty(_rom) ? _rom :
-        _rom.indexOf("_") > 0 ? StringUtils.substringBefore(_rom, "_") + "_" :
-        StringUtils.substring(_rom, 0, 5);
-
     tables.stream()
-        .filter(table -> checkRom(romForFilter, table, checkall))
         .forEach(table -> {
 
           // for debugging purpose and check the matching of a particular table, change id here and add breakpoint
-          //if ("32OBVwzC".equals(table.getId())) {
-          //  boolean stop = true;
-          //}
+          if ("0shM2rcl".equals(table.getId()) || "GNXfb2HW".equals(table.getId())) {
+            boolean stop = true;
+          }
 
-          double dist = getTableDistance(table, _fileName, _tableName, _manuf, _year);
+          double dist = getTableDistance(table, fileName, tableName, manuf, year, rom);
           if (dist < minDistance[0]) {
             found[0] = table;
             minDistance[0] = dist;
@@ -50,31 +47,26 @@ public class TableMatcher {
   /**
    * Take a table display name and a rom and compare it to the given VpsTable
    */
-  public boolean isClose(String _fileName, String _rom, boolean checkall,
-      String _tableName, String _manuf, int _year, VpsTable table) {
-
-    // first compare rom
-    if (!checkRom(_rom, table, checkall)) {
-      return false;
-    }
-
-    double dist = getTableDistance(table, _fileName, _tableName, _manuf, _year);
-    return dist < THRESHOLD_NOTFOUND;
+  public boolean isClose(String fileName, String rom, String tableName, String manuf, int year, VpsTable table) {
+    double dist = getTableDistance(table, fileName, tableName, manuf, year, rom);
+    return dist <= THRESHOLD_CONFIRM;
   }
 
-  private double getTableDistance(VpsTable table, String _fileName, String _tableName, String _manuf, int _year) {
-    String tableName = VpsAutomatcher.cleanChars(table.getName());
+  private double getTableDistance(VpsTable table, String _fileName, String _tableName, String _manuf, int _year, String rom) {
+    String tableName = VpsAutomatcher.cleanTable(table.getName().toLowerCase());
     String manuf = table.getManufacturer();
     String[] altmanufs = getAlternateManuf(manuf);
     int year = table.getYear();
 
     double minDistance = 10000;
-    // return the minimal distance considering alternative manufacturer
-    for (String altmanuf : altmanufs) {
-      double dist = _tableName != null ? tableDistance(_tableName, _manuf, _year, tableName, altmanuf, year)
-          : tableDistance(_fileName, tableName, altmanuf, year);
-      if (dist < minDistance) {
-        minDistance = dist;
+    if (checkRom(rom, table)) {
+      // return the minimal distance considering alternative manufacturer
+      for (String altmanuf : altmanufs) {
+        double dist = _tableName != null ? tableDistance(_tableName, _manuf, _year, tableName, altmanuf.toLowerCase(), year)
+            : tableDistance(_fileName, tableName, altmanuf.toLowerCase(), year);
+        if (dist < minDistance) {
+          minDistance = dist;
+        }
       }
     }
     return minDistance;
@@ -107,11 +99,11 @@ public class TableMatcher {
                                String tableName, String manuf, int year) {
     double dTable = distance(tableName, _tableName) * 1;
     double dManuf = _manuf != null ? distance(_manuf, manuf) : 5;
-    double dYear = _year > 1900 ? (_year == year ? 0 : Math.abs(_year - year) == 1 ? 1 : 5) : 5;
+    double dYear = _year > 1900 ? Math.min(Math.abs(_year - year), 5): 5;
 
     double dist = (1 + dTable)
-        * (1 + dManuf / 35)
-        * (1 + dYear / 5.0);
+        * (1 + dManuf / 35.0)
+        * (1 + dYear / 20.0) - 1;
     return dist;
   }
 
@@ -119,7 +111,7 @@ public class TableMatcher {
     if (StringUtils.isEmpty(str1) && StringUtils.isEmpty(str2)) {
       return 0;
     }
-    else if (StringUtils.equalsIgnoreCase(str1, str2)) {
+    else if (StringUtils.equals(str1, str2)) {
       return 0;
     }
     else if (StringUtils.isEmpty(str1) || StringUtils.isEmpty(str2)) {
@@ -133,13 +125,15 @@ public class TableMatcher {
     //}
     // else
 
-    double ratio = ed.apply(
-        str1.toLowerCase(), 
-        str2.toLowerCase());
+    double ratio = cd.apply(str1, str2);
+
+    double jw = jwd.apply(str1, str2);
+
+    if (true) return ratio * jw; 
 
     // worst ratio, try without spaces
     if (ratio == 1.0) {
-      ratio = ed.apply(
+      ratio = cd.apply(
         str1.replace(" ", "").toLowerCase(), 
         str2.toLowerCase());
     }
@@ -152,20 +146,17 @@ public class TableMatcher {
   /**
    * Take a rom and compare it with the one of the table
    * If rom is empty, verify the table does not use a rom
-   *
    * @return true if this is same rom
    */
-  public boolean checkRom(String rom, VpsTable table, boolean checkall) {
-    if (checkall) {
-      return true;
-    }
-    List<VpsAuthoredUrls> romFiles = table.getRomFiles();
-    // if rom is empty, check the table does not use a rom
-    if (StringUtils.isEmpty(rom)) {
-      return romFiles == null || romFiles.size() == 0;
-    }
-    // else rom should be within the romFiles of the table if present
-    if (romFiles != null && romFiles.size() > 0) {
+  public boolean checkRom(String rom, VpsTable table) {
+
+    //OLE 2025-05-05:
+    // Disconnected the rom filtering as it is too error prone
+    // VPS in many cases does not contain the roms or is missing a variant
+
+    /*List<VpsAuthoredUrls> romFiles = table.getRomFiles();
+    if (StringUtils.isNotEmpty(rom) && (romFiles != null && romFiles.size() > 0)) {
+      // rom should be within the romFiles of the table if present
       boolean found = false;
       for (VpsAuthoredUrls romFile : romFiles) {
         found |= romFile.getVersion() != null
@@ -174,11 +165,9 @@ public class TableMatcher {
             || StringUtils.startsWithIgnoreCase(rom, romFile.getVersion()));
       }
       return found;
-    }
-    // else when rom is expected but the table does not declare any rom 
-    else {
-      return false;
-    }
+    }*/
+    //else
+    return true;
   }
 
   // ------------------------------------------------------