+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/vps/MediaRenamer.java
@@ -10,6 +10,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import de.mephisto.vpin.connectors.vps.VPS;
+import de.mephisto.vpin.connectors.vps.matcher.VpsAutomatcher;
 import de.mephisto.vpin.connectors.vps.model.VpsTable;
 
 /**